Obstacles
When making a flow, randomly assign some tiles cost 1, others cost 10
Enemies can still walk through the 10, but they would prefer not to.
Colour the ground to see it.

Make a flow with f, spawn enemies with e.

Bug
As soon as an enemy enters a tile, it uses its direction
But this has the enemy walk along the edges of tiles mostly.

How do I fix this?
Only update your new direction when you reach the center of a new tile?
Change the direction to pos of neighbour
Get a pos, move towards that position. When really close, update to a enw
position. Would be slow to update if the flow changes underneath the
enemy.

This same bug make enemies stop on the edge of the destination.