Colour Change Propagate
When you change the colour and type of a tile, you can also propagate out
to your neighbours.
Works, but some bugs.

If neighbour range is greater than 1, there are some collisions where
multiple tiles tell the same neighbour to change on the same frame, thus
cycling their TileType twice.

But in the final version, it won't cycle through the tileTypes, it will
set a specific one.
And I think I can do a better version of this system anyway so it doesn't
really matter