Enemy Health, Killcount
Enemies have 5hp

When a bullet kills an enemy, it tells the gun that shot it that it
killed something.
When a Gun gets 5 kills, it gets +1 damage to its bullets.

Bugs
Once the selection was cleared before the tower was built. Some systems
aren't quite in the right order.
Recalculated a new flow field and it didn't work. Was I off the grid? Is
that why? Game seemed to lag like crazy too. Might be connected. Enemies
just continued straight into the distance.