Explosions and Splitting
When bullets end, they can explode or split into 2 copies

If splitting is selected, create 2 copies of the bullet with slightly
offset directions. Destroy the original
If Explosion, create a new circle with an explosion on it. Destroy it
after 0.5s. Destroy the original bullet.

Damage upgrade
Explosion
If the bullet has the explosion ending, the damage boost will go to the
explosion instead of the bullet.
There might be a better way to do this. Like having the collision check if
it will explode. If it will, don't do damage, just spawn the explosion.
And just pass the damage from bullet to bomb.

Bugs fixed
Bullets under background
I think bullets were sometimes under the background because they have the
same z value
Towers not spawning
Figured out the order of systems needed.
It's tower_build_buttons_interactions that sends an event to
build_tower_system which needs to be before grid::clear_selection

Refactors needed
Enemy death. Bullets tell it to die when they kill it. Maybe it should
have a system to moniter its own health
Bullet new() requires a dir, but it gets reset when spawned. That's
confusing
Damage. Bullets should have a damage component that the explosion can also
use instead of having explosion and bullets both checking for collisions
with the enemies