Bomb Tower
Made a bomb tower using mostly the same system as MinesweeperTD

Bomb is a bullet with no collider that moves in an arc and explodes when
it ends.

It's a fairly sloppy implementation, but it works
Some Issues
TowerType and GunType
TowerType makes sense to have multiple, but GunType doesn't. Shotgun,
Pistol, and Bomb aren't that different. They could be the same thing. I
could see an upgrade for the pistol that allows it to shoot multiple
projectiles like a shotgun. And a bomb tower is just a pistol that shoots
different bullets so just swap out the bullet field in the gun.

Enemy Death
Changed to a system the enemy handles instead of the bullet that killed it
cleaning it up