Territory Refresh
Territory Grabber
Handles grabbing terrritory. Allows you to refresh your territory every
10s if enemies capture it (not that they can yet)
Attach this to a tower that can grab territory.
The range seems off by 1. A range of 4 only turns 3 tiles. A range of 2
only turns the 4 adjacent to you. It works, but that's not what I think of/expect when I
set the range to 2.
Might be able to reuse this so enemies can grab territory too. TODO.

Gold
Spawn gold every few seconds
Attached to the NoGun tower. Could attach it to any tower.
Gold doesn't do anything yet. Next todo is to interact with the friendly
territory for more production