Gold Spawner Area
Gold spawns in an area now. Only spawns on friendly territory

Range is now correct.
change_allegiance() for i in 0..range -> i in range 0..=range

Clippy
Cleaned up some clippy warnings

Bugs Fixed
Fixed a bug in get_diamond_ring
I had if x >= 0. I wanted up_x >= 0
I wanted to check if the i32 (which can be negative) was negative before
casting it to usize and to skip if it was negative. But I was accidentally
checking if the usize was negative (whiich it can't be).
This gave me a warning when I built, but I misunderstood it. Clippy gave a
warning with a bit more info so I could figure out what I was actually
doing wrong.