Basic Gun Info
Wanted to expand the functionality of Gun::new()
Made struct BasicGunInfo to store the variables for making a gun. They
have default values and then you can change the things you want to change.
Doesn't contain derived values like the timers.
Now when you create a new Gun, you use Gun::from_basic_gun_info and pass
in the stuff that's unique and use default for the rest.
Basically the same as Gun::new(a, b, c, d, e, ...) but with a struct, I
can hide stuff I don't want to worry about and let the default handle it.