Positional Aiming
Aim at a position, not a direction.
Direction can be calculated when needed.

Bomb/Arc
Made it scale its lifetime and thus speed on the distance. Has a minimum
distance and time and a maximum distance and time.

Bullet::new_arc_complete()
instead of Bullet::new_arc().update_entity().update_parent().update_arc()
Might not be better.

Bullet as a field and component
I thought it would be handy to just store a Bullet component in a field on
the gun and then I could just directly insert it when I spawned a bullet.
But there are some issues like the gun not knowing the direction it will
shoot a bullet until it is created.
So maybe I should change to a BulletInfo and BulletComponent structure
where you pass the info into the constructor of the bullet component plus
the other info needed at the time of creation like the direction.