Stacking Gold
Tiles can have gold on them.
When you swap, if they both had gold on them, stack the gold together on
the tile you are holding.
When a gold spawner spawns gold, now it gets stored on the tile. Before it
was just an animation that did nothing.

Swap.rs
Moved swap logic from grid to its own file