Swap Refactor
Moved the steps of the swapping into their own systems
When a swap happens, send out an event. Each system listens to the event
and reacts how they want.

I don't know if this makes things less coupled or easier to read.

Appears to work. Should be parrallel. The event sends out the entities so
it doesn't rely on using coords to grab the entity from the grid which
would be different after the grid swaps.